Move GlobalSetup to MultiRegionSetup
Move the "global" setup as it is now only used for multi-region tests.
