--- JustSaying.IntegrationTests/GlobalSetupCollection.cs
@@ -1,9 +1,0 @@
-using Xunit;
-
-namespace JustSaying.IntegrationTests
-{
-    [CollectionDefinition(GlobalSetup.CollectionName, DisableParallelization = true)]
-    public class GlobalSetupCollection : ICollectionFixture<GlobalSetup>
-    {
-    }
-}
