+++ JustSaying.IntegrationTests/JustSayingFluently/MultiRegion/MultiRegionSetupCollection.cs
@@ -1,0 +1,9 @@
+using Xunit;
+
+namespace JustSaying.IntegrationTests.JustSayingFluently.MultiRegion
+{
+    [CollectionDefinition(MultiRegionSetup.CollectionName, DisableParallelization = true)]
+    public class MultiRegionSetupCollection : ICollectionFixture<MultiRegionSetup>
+    {
+    }
+}
